fix(install/desktop): only sudo rm if necessary

+++ install/desktop/set-app-grid.sh
@@ -1,2 +1,4 @@
 # App doesn't do anything when started from the app grid
-sudo rm -rf /var/lib/flatpak/exports/share/applications/org.flameshot.Flameshot.desktop
+if [ -f "/var/lib/flatpak/exports/share/applications/org.flameshot.Flameshot.desktop" ]; then
+  sudo rm -rf /var/lib/flatpak/exports/share/applications/org.flameshot.Flameshot.desktop
+fi
